Replace all functionality in core.py
